Refactor: move NewCard class to models dir
