+++ src/models/NewCard.ts
@@ -1,0 +1,31 @@
+import { FieldValue, serverTimestamp } from "firebase/firestore";
+import { nanoid } from "nanoid";
+import { ICard, cardStatus } from "../store/cardStore";
+
+export class NewCard implements ICard {
+  id: string;
+  content: string;
+  tags: string[];
+  status: cardStatus;
+  isArchived: boolean = false;
+  isTransformed: boolean = false;
+  isImportant: boolean = false;
+  serverTimestamp: FieldValue;
+  createdTime: string;
+  updatedTime: string;
+  dueDate: string | null;
+  reminderStartDate: number | null = null;
+  reminderEndDate: number | null = null;
+
+  constructor(content: string, tags: string[], status: cardStatus) {
+    const currentDate = new Date().toISOString();
+    this.id = nanoid();
+    this.content = content;
+    this.tags = tags[0] ? tags : [];
+    this.status = status;
+    this.serverTimestamp = serverTimestamp();
+    this.createdTime = currentDate;
+    this.updatedTime = currentDate;
+    this.dueDate = status === "todo" ? currentDate : null;
+  }
+}
