+++ src/store/cardStore.ts
@@ -7,13 +7,12 @@ import {
   getDoc,
   onSnapshot,
   query,
-  serverTimestamp,
   setDoc,
   where,
 } from "firebase/firestore";
 import { makeAutoObservable, runInAction } from "mobx";
-import { nanoid } from "nanoid";
 import { db } from "../config/firebase";
+import { NewCard } from "../models/NewCard";
 import { cardUtils } from "../utils/cardUtils";
 import { cookie } from "../utils/cookie";
 import { debounceCardsOrderList } from "../utils/debounce";
@@ -67,34 +66,6 @@ interface IFirebaseService {
 /* ===============================
 ===========  NewCard  ============
 =============================== */
-
-export class NewCard implements ICard {
-  id: string;
-  content: string;
-  tags: string[];
-  status: cardStatus;
-  isArchived: boolean = false;
-  isTransformed: boolean = false;
-  isImportant: boolean = false;
-  serverTimestamp: FieldValue;
-  createdTime: string;
-  updatedTime: string;
-  dueDate: string | null;
-  reminderStartDate: number | null = null;
-  reminderEndDate: number | null = null;
-
-  constructor(content: string, tags: string[], status: cardStatus) {
-    const currentDate = new Date().toISOString();
-    this.id = nanoid();
-    this.content = content;
-    this.tags = tags[0] ? tags : [];
-    this.status = status;
-    this.serverTimestamp = serverTimestamp();
-    this.createdTime = currentDate;
-    this.updatedTime = currentDate;
-    this.dueDate = status === "todo" ? currentDate : null;
-  }
-}
 
 export class NewNote extends NewCard {
   noteTitle: string = "";
